Return absolute paths in breakpoint responses.
When debug symbols have relative paths, lldb's BreakpointCreateByLocation returns the relative path (even when target.source-map is set). VSCode will treat the path as relative to the workspace folder, and if it is not the breakpoint will show disabled and will disapear from the file (lldb will stil break, but there is no way to unset it). Also, clicking on the breakpoint will prompt to create the file, as VSCode cannot find it.