Support DAP disassemble request
CodeLLDB currently suppports a custom disassembly view and provides
disassembly as "source" when debugging into objects with no sourceline
info.

DAP now also has a `disassemble` request which, given a memory refernce
from the stack trace, produces a set number of instructions from that
address.

This API is awkward and annoying, but it's simple to implement based on
the existing DisassembledRange.

WIP.
 - we don't return the exact number of instructions
 - we don't populate a lot of the optional fields
 - my-first-rust(TM)
 - no tests yet