Allow dumping of memory based on register value
It's pretty common when lookoing at disassembly to see RSP+some offset.
If you want to dump the memory at RSP, you naturally go to the scopes
screen, find the RSP register and dump it. But we don't currently set
any memory reference for registers.

Solution: For a register-type value, assume that it is a memory address
and pass as its value as a memory reference. This might be bad if the
user tries to dump 0 or some random int, but what can we do?

TODO:
 - can we check if the resulting value is a valid load address in a
   mapped segment ?
 - can we do the same for "any" literal value, allowing users to just
   type in an address and dump its memory (there's no provision for this
   in the protocol, unbelievably)