Evaluate request: return memory reference
This implements 2 things:

First, we now return memoory references for all members of an evaluate
request's response. This allows for dumping of memory around the results
of mundane evaluation, such as variables, structs, etc.

Further, any evaluation result which is itself a pointer value gets the
contained address as its memory address (as opposed to the load address
of the pointer itslef). This in particular allows for dumping of
arbitrary memory addresses by simply casting them to (void*) or some
equivalent pointer type. This is consistent with other debug adapters.