Add full support for data breakpoints
Currently, we only provode support for data breakpoints where the
request is for a child of some container using variables_reference.
However, the spec includes support for:
- an arbitrary expression, and
- an arbitrary address, and
- an optional size in bytes.
The later two are supported via a capability "supportsDataBreakpointBytes".
We add support for all of these options.

It's fairly trivial to evaluate an expression and do esentially what we
do now for children (find the load address, check its size). But we can
also accept a size, so long as it's valid watchpoint size. Similarly for
addresses, if 'asAddress' is supplied we can just try to use it. It will
probably fail if the address is actually invalid, but that's sort of the
point.