Allow any size for Watchpoints
Current released versions of lldb have seeming partial support for
"large" watchpoints. That is they break a user requested watchpoint and
size into specific minimal set of working hardware watchpoints supported
by the target.

Full support for this is in progress in master and works well, allowing
to watch any address that's part of some larger structure. In any case,
CodeLLDB rejecting a watch due to its size is swimming against the tide
and by removing the size restriction we can take advantage of whatever
support LLDB adds in the future versions without code changes.

The drawback is that we get a slightly less precise error
"Setting one of the watchpoint resources failed" in the case where the
size is not currently supported.